Refactor core blocks to have save and transforms in their own files (part 2) (#14899)
* Refactor core blocks to have save and transforms in their own files (part 2)

* Fix whitespace in save function from  Gallery
