+++ packages/block-library/src/heading/save.js
@@ -1,0 +1,17 @@
+/**
+ * WordPress dependencies
+ */
+import { RichText } from '@wordpress/block-editor';
+
+export default function save( { attributes } ) {
+	const { align, level, content } = attributes;
+	const tagName = 'h' + level;
+
+	return (
+		<RichText.Content
+			tagName={ tagName }
+			style={ { textAlign: align } }
+			value={ content }
+		/>
+	);
+}
