+++ packages/block-library/src/latest-comments/index.js
@@ -13,19 +13,13 @@ export const name = 'core/latest-comments';
 
 export const settings = {
 	title: __( 'Latest Comments' ),
-
 	description: __( 'Display a list of your most recent comments.' ),
-
 	icon,
-
 	category: 'widgets',
-
 	keywords: [ __( 'recent comments' ) ],
-
 	supports: {
 		align: true,
 		html: false,
 	},
-
 	edit,
 };
