+++ packages/block-library/src/latest-posts/index.js
@@ -13,19 +13,13 @@ export const name = 'core/latest-posts';
 
 export const settings = {
 	title: __( 'Latest Posts' ),
-
 	description: __( 'Display a list of your most recent posts.' ),
-
 	icon,
-
 	category: 'widgets',
-
 	keywords: [ __( 'recent posts' ) ],
-
 	supports: {
 		align: true,
 		html: false,
 	},
-
 	edit,
 };
