+++ packages/block-library/src/legacy-widget/index.js
@@ -13,16 +13,11 @@ export const name = 'core/legacy-widget';
 
 export const settings = {
 	title: __( 'Legacy Widget (Experimental)' ),
-
 	description: __( 'Display a legacy widget.' ),
-
 	icon,
-
 	category: 'widgets',
-
 	supports: {
 		html: false,
 	},
-
 	edit,
 };
