+++ packages/block-library/src/list/index.js
@@ -7,13 +7,7 @@ import { omit } from 'lodash';
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import {
-	createBlock,
-	getPhrasingContentSchema,
-	getBlockAttributes,
-} from '@wordpress/blocks';
 import { RichText } from '@wordpress/block-editor';
-import { replace, join, split, create, toHTMLString, __UNSTABLE_LINE_SEPARATOR } from '@wordpress/rich-text';
 
 /**
  * Internal dependencies
@@ -21,27 +15,12 @@ import { replace, join, split, create, toHTMLString, __UNSTABLE_LINE_SEPARATOR }
 import edit from './edit';
 import icon from './icon';
 import metadata from './block.json';
+import save from './save';
+import transforms from './transforms';
 
 const { name, attributes: schema } = metadata;
 
 export { metadata, name };
-
-const listContentSchema = {
-	...getPhrasingContentSchema(),
-	ul: {},
-	ol: { attributes: [ 'type' ] },
-};
-
-// Recursion is needed.
-// Possible: ul > li > ul.
-// Impossible: ul > ul.
-[ 'ul', 'ol' ].forEach( ( tag ) => {
-	listContentSchema[ tag ].children = {
-		li: {
-			children: listContentSchema,
-		},
-	};
-} );
 
 const supports = {
 	className: false,
@@ -52,118 +31,8 @@ export const settings = {
 	description: __( 'Create a bulleted or numbered list.' ),
 	icon,
 	keywords: [ __( 'bullet list' ), __( 'ordered list' ), __( 'numbered list' ) ],
-
 	supports,
-
-	transforms: {
-		from: [
-			{
-				type: 'block',
-				isMultiBlock: true,
-				blocks: [ 'core/paragraph' ],
-				transform: ( blockAttributes ) => {
-					return createBlock( 'core/list', {
-						values: toHTMLString( {
-							value: join( blockAttributes.map( ( { content } ) => {
-								const value = create( { html: content } );
-
-								if ( blockAttributes.length > 1 ) {
-									return value;
-								}
-
-								// When converting only one block, transform
-								// every line to a list item.
-								return replace( value, /\n/g, __UNSTABLE_LINE_SEPARATOR );
-							} ), __UNSTABLE_LINE_SEPARATOR ),
-							multilineTag: 'li',
-						} ),
-					} );
-				},
-			},
-			{
-				type: 'block',
-				blocks: [ 'core/quote' ],
-				transform: ( { value } ) => {
-					return createBlock( 'core/list', {
-						values: toHTMLString( {
-							value: create( { html: value, multilineTag: 'p' } ),
-							multilineTag: 'li',
-						} ),
-					} );
-				},
-			},
-			{
-				type: 'raw',
-				selector: 'ol,ul',
-				schema: {
-					ol: listContentSchema.ol,
-					ul: listContentSchema.ul,
-				},
-				transform( node ) {
-					return createBlock( 'core/list', {
-						...getBlockAttributes(
-							'core/list',
-							node.outerHTML
-						),
-						ordered: node.nodeName === 'OL',
-					} );
-				},
-			},
-			...[ '*', '-' ].map( ( prefix ) => ( {
-				type: 'prefix',
-				prefix,
-				transform( content ) {
-					return createBlock( 'core/list', {
-						values: `<li>${ content }</li>`,
-					} );
-				},
-			} ) ),
-			...[ '1.', '1)' ].map( ( prefix ) => ( {
-				type: 'prefix',
-				prefix,
-				transform( content ) {
-					return createBlock( 'core/list', {
-						ordered: true,
-						values: `<li>${ content }</li>`,
-					} );
-				},
-			} ) ),
-		],
-		to: [
-			{
-				type: 'block',
-				blocks: [ 'core/paragraph' ],
-				transform: ( { values } ) =>
-					split( create( {
-						html: values,
-						multilineTag: 'li',
-						multilineWrapperTags: [ 'ul', 'ol' ],
-					} ), __UNSTABLE_LINE_SEPARATOR )
-						.map( ( piece ) =>
-							createBlock( 'core/paragraph', {
-								content: toHTMLString( { value: piece } ),
-							} )
-						),
-			},
-			{
-				type: 'block',
-				blocks: [ 'core/quote' ],
-				transform: ( { values } ) => {
-					return createBlock( 'core/quote', {
-						value: toHTMLString( {
-							value: create( {
-								html: values,
-								multilineTag: 'li',
-								multilineWrapperTags: [ 'ul', 'ol' ],
-							} ),
-							multilineTag: 'p',
-						} ),
-					} );
-				},
-			},
-		],
-	},
-
+	transforms,
 	deprecated: [
 		{
 			supports,
@@ -197,7 +66,6 @@ export const settings = {
 			},
 		},
 	],
-
 	merge( attributes, attributesToMerge ) {
 		const { values } = attributesToMerge;
 
@@ -210,15 +78,6 @@ export const settings = {
 			values: attributes.values + values,
 		};
 	},
-
 	edit,
-
-	save( { attributes } ) {
-		const { ordered, values } = attributes;
-		const tagName = ordered ? 'ol' : 'ul';
-
-		return (
-			<RichText.Content tagName={ tagName } value={ values } multiline="li" />
-		);
-	},
+	save,
 };
